src/requirements.txt
python-decouple
pandas
numpy
matplotlib
pathlib
opencv-python
scikit-learn
tensorflow
icrawler
Google-Images-Search